feat: API 개발 완료

+++ src/main/java/com/example/community/dto/post/PostCreateResponseDto.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class PostCreateResponseDto {
-    private Integer postId;
+    private Long postId;
 }
